Move logic from CrosswordField into the useCrosswordField custom hook
